Re-added code to use cached script is present

+++ jenkins/pipelines/QE/dotnet/prepare_env.sh
@@ -7,7 +7,9 @@ SCRIPT_DIR=$( cd -- "$( dirname -- "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )
 
 PREPARE_DOTNET_SCRIPT="$SCRIPT_DIR/prepare_dotnet.sh"
 
-echo "Downloading prepare_dotnet.sh..."
-curl -L https://raw.githubusercontent.com/couchbaselabs/couchbase-mobile-tools/refs/heads/master/dotnet_testing_env/prepare_dotnet_new.sh -o $PREPARE_DOTNET_SCRIPT
+if [ ! -f $PREPARE_DOTNET_SCRIPT ]; then
+    echo "Downloading prepare_dotnet.sh..."
+    curl -L https://raw.githubusercontent.com/couchbaselabs/couchbase-mobile-tools/refs/heads/master/dotnet_testing_env/prepare_dotnet_new.sh -o $PREPARE_DOTNET_SCRIPT
+fi
 
 source $PREPARE_DOTNET_SCRIPT
